+++ utils/resizer/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - [`1d9a662`](https://github.com/WPMedia/feed-components/commit/1d9a662eab4b33265d14b9c7079904569e583c63) [#69](https://github.com/WPMedia/feed-components/pull/69) Thanks [@tbrick855](https://github.com/tbrick855)! - added width and height parameters
 
+## 0.1.2
+
 ## 0.1.1
 
 ### Patch Changes